司机与列车组成树结构

+++ nick-api/src/main/java/com/nick/api/mapper/MetroInfoMapper.java
@@ -11,7 +11,7 @@ import org.apache.ibatis.annotations.Mapper;
  */
 @Mapper
 public interface MetroInfoMapper  {
-     List<MetroInfo> getAllByMetroId(Long id);
+     List<MetroInfo> getAllByMetroId(Integer id);
 }
